Move build script and run script to official plugins
