+++ lerna.json
@@ -1,8 +1,8 @@
 {
   "version": "0.0.0",
   "packages": [
-    "packages/*",
-    "packages/@snowpack/*"
+    "packages/@snowpack/*",
+    "packages/*"
   ],
   "npmClient": "yarn",
   "useWorkspaces": true
