+++ packages/@snowpack/app-scripts-preact/package.json
@@ -11,7 +11,7 @@
     "@babel/plugin-syntax-import-meta": "^7.8.3",
     "@babel/preset-env": "^7.9.5",
     "@babel/preset-react": "^7.9.4",
-    "@babel/preset-typescript": "^7.9.0",
+    "@babel/preset-typescript": "^7.10.4",
     "@prefresh/snowpack": "0.1.0",
     "@snowpack/plugin-babel": "^1.4.0",
     "@snowpack/plugin-dotenv": "^1.6.0-alpha.0",
