+++ packages/@snowpack/app-template-lit-element-typescript/package.json
@@ -23,7 +23,7 @@
     "@snowpack/app-scripts-lit-element": "^1.6.0-alpha.0",
     "prettier": "^2.0.5",
     "snowpack": "^2.6.4",
-    "typescript": "^3.8.3"
+    "typescript": "^3.9.7"
   },
   "gitHead": "795f4311d79a70cc9f19f21b512b7f8675d73f17"
 }
