+++ packages/@snowpack/app-template-react-typescript/package.json
@@ -28,7 +28,7 @@
     "jest": "^25.4.0",
     "prettier": "^2.0.5",
     "snowpack": "^2.6.4",
-    "typescript": "^3.8.0"
+    "typescript": "^3.9.7"
   },
   "gitHead": "795f4311d79a70cc9f19f21b512b7f8675d73f17"
 }
