+++ packages/@snowpack/app-template-svelte-typescript/package.json
@@ -25,7 +25,7 @@
     "snowpack": "^2.6.4",
     "svelte-check": "^0.1.53",
     "svelte-preprocess": "^4.0.8",
-    "typescript": "^3.9.6"
+    "typescript": "^3.9.7"
   },
   "gitHead": "795f4311d79a70cc9f19f21b512b7f8675d73f17"
 }
